+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/search/DDMFormInstanceRecordBatchReindexer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.internal.search.util;
+package com.liferay.dynamic.data.mapping.internal.search;
 
 /**
  * @author Rafael Praxedes
